+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 3.0 (Jan, 2026)
 
+### Added
+
+- Added tags to small view.
+
 
 ## 2.9 (Jan, 2026)